# lib/api-zod/src/generated/types/activityEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Brandex DMS API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityEntry {
  id: number;
  entityType: string;
  entityId: number;
  action: string;
  /** @nullable */
  details?: string | null;
  createdAt: Date;
}