# lib/api-zod/src/generated/types/case.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Brandex DMS API
 * OpenAPI spec version: 0.1.0
 */

export interface Case {
  id: number;
  clientId: number;
  /** @nullable */
  clientCode?: string | null;
  /** @nullable */
  clientName?: string | null;
  /** 6-digit trademark number */
  trademarkNumber: string;
  trademarkClass: number;
  /**
     * @minimum 1
     * @maximum 45
     */
  stage: number;
  /** @nullable */
  requiredDate?: Date | null;
  /** @nullable */
  description?: string | null;
  due?: number;
  received?: number;
  total?: number;
  createdAt: Date;
  updatedAt?: Date;
}