# lib/api-zod/src/generated/types/clientInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Brandex DMS API
 * OpenAPI spec version: 0.1.0
 */

export interface ClientInput {
  clientCode: string;
  name: string;
  city?: string;
  email?: string;
  imageUrl?: string;
  notes?: string;
}