# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Brandex DMS API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEntry } from './activityEntry';
import type { DashboardSummaryCasesByStageItem } from './dashboardSummaryCasesByStageItem';

export interface DashboardSummary {
  totalClients: number;
  totalCases: number;
  totalDue: number;
  totalReceived: number;
  balance: number;
  casesByStage: DashboardSummaryCasesByStageItem[];
  recentActivity: ActivityEntry[];
}