# lib/api-zod/src/generated/types/payment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Brandex DMS API
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentType } from './paymentType';

export interface Payment {
  id: number;
  clientId: number;
  /** @nullable */
  caseId?: number | null;
  /** @nullable */
  trademarkNumber?: string | null;
  /** @nullable */
  clientName?: string | null;
  type: PaymentType;
  amount: number;
  date: Date;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}